refactor: split collections panel ui
